feat : Car 에 winner column 추가 후 코드 변경
- table 스펙 변경에 따라 Dao 수정
- Dao 수정에 따른 Serivce & test 수정

+++ src/test/java/racingcar/service/CarServiceIntegrationTest.java
@@ -25,7 +25,7 @@ class CarServiceIntegrationTest {
                 RacingGame.readyToRacingGame("a,b,c",
                                              new RandomNumberGenerator(),
                                              trialCount);
-        final int savedRaceResultId = 3;
+        final Long savedRaceResultId = 3L;
 
         //when & then
         assertDoesNotThrow(() -> carService.registerCars(racingGame, savedRaceResultId));
